directio: Use shared map for directio
We use a shared map to make direct io safe if fork is invoked in another
thread concurrently with the direct io.

Using private maps with direct io can cause data corruption and
undefined behavior in the parent or the child processes. This
restriction does not apply to memory buffer created with MAP_SHARED.
See open(2) for more info.

Change-Id: Ib0bf88f0cb9bd6fd37e4c9b838aed75330447176
Signed-off-by: Nir Soffer <nsoffer@redhat.com>