directio: Add _directio extension module
Python cannot do directio read properly due to misdesign of built-in
fileobject's readinto(). Add an extension module providing an aligned
buffer and efficient way to read data from a file descriptor or copy
data from strings or other buffers.

The buffer implements the buffer interface so it can be written to
another file object or to a file descriptor without additional copy.

Change-Id: I979a6c2dc00271a8f71b35e56b29ac3c2b0f9f4a
Signed-off-by: Nir Soffer <nsoffer@redhat.com>