errors: Introduce the errors module
Add new errors module for keeping imaged errors. The first error is
PartialContent, raised when available size is shorter then the requested
size.  Add tests verifying that the new error is raised when it should.

The new tests reveal that the reading loop was not checking the partial
content correctly when a file is not aligned to 512 bytes. The last read
returned partial content leaving the file offset at the end of the file.
The next read fails because the file offset is not aligned. Now we check
this state and fail wit the new PartialContent exception.

Change-Id: I9066ec43c71e6ce873abd16cac68c5d283694b3d
Signed-off-by: Nir Soffer <nsoffer@redhat.com>