bench: Abort progress when operation fail
Previously the operation was run in a background thread, and progress
run in the main thread. If the thread raised, progress continued to
watch the dead thread forever.

Run the operation in the main thread, and the progress in a background
thread, so failing operation stops the process.

Example failure:

$ ./bench send /var/tmp/image 10000 >/dev/null
Traceback (most recent call last):
  File "./bench", line 78, in <module>
    args.command(args)
  File "./bench", line 24, in send
    run(op)
  File "./bench", line 35, in run
    op.run()
  File "/home/nsoffer/src/vdsm-imaged/imaged/directio.py", line 63, in run
    self._send_chunk(src, buf)
  File "/home/nsoffer/src/vdsm-imaged/imaged/directio.py", line 75, in _send_chunk
    raise errors.PartialContent(self.size, self.done)
imaged.errors.PartialContent: Requested 10485760000 bytes, available 8388608000 bytes

Change-Id: I74870c53dac8e02fbb582424a86277a8c15a3596
Signed-off-by: Nir Soffer <nsoffer@redhat.com>