test: HTTP GET, OPTIONS unit tests for server module
Add unit tests for the server module, which indirectly tests the
image_handler module as well.  This set of tests covers HTTP GET and
OPTIONS.

Some supplementary files are also added, providing a valid configuration
for the server to run during testing along with a pre-generated auth
ticket acceptable as input to the server.  (This test ticket will be
valid for approximately 30 years.)

This also requires the python-requests-mock extension module:

The proxy uses python-requests to send http[s] requests to imaged, and
in order to fully test the functionality it expects certain responses
from imaged, otherwise it will fail early preventing us from testing all
code paths. Requests-mock is designed as an adapter for python-requests
which crafts configurable responses to queries so that unit testing code
using python-requests doesn't require the availability of an external
network resource.

An alternative, HTTPretty, was ruled out because its monkey-patching
around the socket layer caused problems with webob, which prevented the
image server from starting up.  Requests-mock doesn't interfere with
webob's access to the network, while allowing us to emulate responses
from the vdsm-image-daemon.

For more documentation on python-requests-mock, see:
  http://requests-mock.readthedocs.org/en/latest/overview.html

Change-Id: I8cdf9bfa08cc488033277f90758764b3b0c71c81
Signed-off-by: Greg Padgett <gpadgett@redhat.com>