server: Accept ticket timeout instead of expire timestamp
Previously we used to accept an expire timestamp, which may break if
engine and host clocks are not synchronized. Also, this force usage of
time.time(), which may change backward or forward, invalidating tickets
or giving them more time.

We accept now a timeout in seconds, and compute the expire time when a
ticket is added. When a ticket is extended, we add extend the expire
time by the timeout.

Since we create the expire time, we can use now monotonic clock source
so ticket are not effected now by system time changes.

Change-Id: I13f5de9ac99827d937b9d2899e96cf86f12cabe4
Signed-off-by: Nir Soffer <nsoffer@redhat.com>