server: Return Content-Length header
If a webob caller assigns a stream as the response body, webob will then
clear the Content-Length header for the outgoing response and will not
create one automatically, expecting it to be supplied by the caller.

To compensate, manually copy the Content-Length as returned from
ovirt-image-daemon.  Add unit tests to ensure it works.

Change-Id: I79fa94112c0509a35c512b0b8d2721b2bea5ac4c
Signed-off-by: Greg Padgett <gpadgett@redhat.com>