packaging: move image-daemon to a dedicated folder.
As a decision was made to move image-proxy and image-daemon under the
same repository, this patch moves image-daemon related files to its
specific directory.

Once this change is done, the next phase will be to start migrating code
from the image-proxy project into this one. The proxy code will also
reside in its own directory.

After that, common code and setup scripts will be made for easy setup.

Change-Id: I49a58d933272ee17e27666ba4a7ff55300dcadfe
Signed-off-by: Amit Aviram <aaviram@redhat.com>