common: extract daemon's Application class
As the Application class is implemented almost the same in both
daemon and proxy projects, we move it to a common project which
will be imported to each of the projects.

As a start, only the daemon will use it because the proxy made
some enhancements to it that needs to be gradually imported in
next patches.

Change-Id: I4853441307179fe5f8c1fb7fbbfa1d15257376b0
Signed-off-by: Amit Aviram <aaviram@redhat.com>