daemon, tests: Make tests to support updated version of py.test
Updated versions of py.test prints the parameterized test
values without using its __str__() method.

That fact is causing the param(str) class to be redundant, and makes the
tests to load the shell heavilly.

This patch changes the approach of representing the data sent to the
tests, to use the "ids" callable argument provided by py.test API, with
a "head" function that slices the data, like param class did before.

Change-Id: Ib43bbad1d8dab9bd42f6432c8f08d168142bc2ca
Signed-off-by: Amit Aviram <aaviram@redhat.com>