download: Add GET request handler
The best way to download a file using webob is to use an app_iter
providing the download data. Change directio.Send to support iteration
so we can send it to the response.

Override wsgiref.simple_server.ServerHandler to work with buffer objects
and override WSGIRequestHandler.handle to use our fixed ServerHandler.
These overrides are horrible but I could not find a better way.

Notes:

- We require a valid Range header - we probably need to make it optional
  and return the entire image (as specified in the ticket) if a range
  was not specified.
- Need to handle partial Range (e.g 0-*). Currently we fail with
  HTTPInternalServerError.
- Need to fail nicer when range is invalid

Change-Id: Ida7d80d494a5da61d67cd7f47073955e8fc444e6
Signed-off-by: Nir Soffer <nsoffer@redhat.com>