Proxy: use wsgicors for CORS support.
Currently the proxy uses its own implementation of CORS handling. this
patch replaces the current implementation with the wsgicors module, for
making the code less cumbersome, without re-implementing logic that
already exists.

Note, that the tests are changed as well, as when making a preflight
with the OPTIONS method, other headers are supposed to return than when
making any other call (GET/POST/PATCH), and wsgicors acts that way.
(differently than the original tests)

Change-Id: Idb1d4ba4de0ec441a2fe4cd5235655e32f591ec0
Signed-off-by: Amit Aviram <aaviram@redhat.com>