proxy: Update code, comments, and configuration to use "ticket"
Due to prior changes in the ticket format, the proxy code used both the
terms "token" and "ticket" somewhat interchangably.

To make the code easier to follow, this patch standardizes on the
more commonly-used term "ticket".  In addition, it changes an old
configuration setting name to more accurately reflect its purpose,
and updates a comment to not refer to a prior ticket format.

Change-Id: I24b9c492951b2372972bc201bc56c315f8ab0de9
Signed-off-by: Greg Padgett <gpadgett@redhat.com>