packaging: Commit to ovirt build standards
http://ovirt-infra-docs.readthedocs.io/en/latest/CI/Build_and_test_standards.html
describes standard artifacts that oVirt projects should have for
it to be involved in the standard build process.

This patch adds what's described in the instructions in the link above
to the ovirt-imageio project.

2 tests fail for unknown reason when running in mock; these tests are
marked now as "noci", and we skip these tests using py.test -m feature.
We need to make this funcionality available via the main makefile later.

Change-Id: Idff6a5015d25e1c9a7ee1ed6d4abc5d584d48b32
Signed-off-by: Amit Aviram <aaviram@redhat.com>