proxy: make test to know common automatically.
Currently, the common project needs to be installed/copied locally/added
to syspath in order to run the proxy's tests. this patch makes it easier
to test this project, adding the common project path to syspath
automatically. This will make the developer only calling only "py.test"
in order to test the proxy.

Change-Id: I168d6422e61ac600df27ccd025e685755b5471bb
Signed-off-by: Amit Aviram <aaviram@redhat.com>