proxy: Installation
Add ovirt-engine-setup configuration and plugins.

This follows a similar approach to the websocket proxy setup.  During
engine-setup, it will ask if you wish to configure the Image I/O Proxy.
If yes, and you are on the same host as engine, everything will be done
automatically.  If you are not installing it on the same host as engine,
you will be prompted to follow a short sequence of steps to perform the
PKI setup.

In addition, this will install the firewall settings (if desired) and
configure systemd to start the proxy service.

Change-Id: Id9ed1c53f7150440b0276d96d39a77ff4e857f8a
Signed-off-by: Greg Padgett <gpadgett@redhat.com>