proxy: Update README and licensing information
The supplemental README and licensing information was out of date and/or
wrong.  Sychronize this information with the proper state of the project
to include proper build/test instructions in the readme.

In addition, clarify the license for the proxy (GPLv2) and setup plugin
(ASL 2.0).  The license for the setup plugin is ASL rather than GPL
because this is the precendent for engine-setup, its plugins, and the
OTOPI installation system.

Change-Id: I16b8e4330d3ff8fa91cb131d7fddd2d815077367
Signed-off-by: Greg Padgett <gpadgett@redhat.com>