tests: Use tox to run the tests
tox make it easy to run the tests with different versions of python and
avoid dependencies on old development tools on the platform.

systemd-daemon is not available via pip, so we have use the local
installed package. This require the sitepackages = True option.
When using sitepackages, tox may find and use the local installed
py.test, and skip installing py.test. The whitelist_externals option
silence tox warning about that.

We should improve this later by separating the startup code requiring
systemd-daemon from tested code, so we can disable sitepackages.

Change-Id: I83c88c165b56ce6a5816058c610c495064655458
Signed-off-by: Nir Soffer <nsoffer@redhat.com>