setup: Store address of Image Proxy in database
This patch was migrated from the ovirt-engine project, originally
written by gpadgett, with changes to fit imageio project.

The original patch:
https://gerrit.ovirt.org/#/c/58591

Description:
The FQDN of the host running the Image I/O Proxy is needed in the
vdc_options table so that this can be passed to the webadmin client.
Without the proper hostname (which matches the hostname used during PKI
setup), the proxy cannot validate the authenticity of the ticket used
during the upload process.

Change-Id: I87353966df26841c2f9d584a0ff16193374d8b6a
Bug-Url: https://bugzilla.redhat.com/1342322
Signed-off-by: Amit Aviram <aaviram@redhat.com>