Proxy: Verify daemon's CA signature
Currently the communication between the proxy and the daemon is using
SSL, but the proxy doesn't verify the signature of the CA sent in the
daemon's certificate.

As the daemon uses VDSM's certificate, it is signed by the CA managed by
the engine. therefor, when getting this certificate- compare it to the
CA certificate installed in the proxy.

Change-Id: Ia8900a08d9d3ae8298e2f015635ac0b1bbac13fb
Signed-off-by: Amit Aviram <aaviram@redhat.com>