daemon: Handle missing or invalid Content-Length
We did not check for missing or invalid Content-Length header
(request.content_length is None). Such request was failing with:

    TypeError: unsupported operand type(s) for +: 'int' and 'NoneType'

We also did not check for negative value, failing later in mmap.mmap()
with:

    error: [Errno 22] Invalid argument

Now we return a Bad Request error with a clear description of the
problem instead of Internal Server Error.

Change-Id: I503356a8bfc02145c81ba6d4f74577e48fe189f7
Bug-Url: https://bugzilla.redhat.com/1357269
Signed-off-by: Nir Soffer <nsoffer@redhat.com>