spec: Require vdsm since we depend on vdsm user
ovirt-imageio-daemon does not know anything about vdsm, but it run as
vdsm user to access vdsm images and pki files. When installing on a
machine without the vdsm user, the log directory is created as
root:root, and the daemon will fail later to start, trying to access to
write to the log.

Add Requires(pre) so vdsm must be installed before ovirt-imageio-daemon.

Change-Id: I6cc75c0ff991797aeb2734e9fd4fe5b9a378bc61
Bug-Url: https://bugzilla.redhat.com/1401901
Signed-off-by: Nir Soffer <nsoffer@redhat.com>