common: Move CappedStream to web module
The CappedStream class was used by the proxy for reading request data.
This code was copied from the web without proper review, or tests, and
the result is over complicated and broken.

Changes:

- The class was simplified to implement __iter__ without a separate
  iterator class.
- The read() method use now the standard read(n=None) interface that
  file-like object should implement.
- Move class to web module, this is infrastructure, not application
  code.
- Add the missing tests

Note: One of the tests is marked as expected failure since the capped
stream does not count bytes properly yet.

Change-Id: I6e7ba738e28dcf931c67b02f64c0a29b7e524b4d
Signed-off-by: Nir Soffer <nsoffer@redhat.com>