directio: Prepare for reporting download progress
When a request is done, the response and the app_iter are closed; at
this point the entire response was sent, and the request can be
considered as completed.

Add Operation.close property and Operation.close() method, so an
operation is marked as closed when finishing a request. Also mark
operation as closed when running it, both on success and errors.

This change allows detection of a completed download by attaching the
operations to the ticket, and exposing this information via the /tickets
service.

If the client is still downloading, we will have:

- unclosed operation with some offset and done values

A completed download performed with one request will have:

- closed operation with offset=0 and done=ticket["size"]

If the download was paused and resumed, we will have:

- closed operation with offset=0 and done=N
- closed operation with offset=N and done=(ticket["size"] - N)

If the download was paused and resumed several times, we will have more
operations.

If a user agent is trying to download concurrently, we may have several
operations with different offsets and sizes. We may detect that some
data was not downloaded yet by looking at the operation ranges. Here is
a possible example with 4 concurrent operations on 4g image:

- closed operation with offset=0 and done=1073741824
- closed operation with offset=1073741824 and done=1073741824
- closed operation with offset=2147483648 and done=1073741824
- closed operation with offset=3221225472 and done=1073741824

The data sent from the daemon may have lost in the way to the user
agent, so we cannot tell if the download was completed. When a user
agent is downloading images, we are not in the loop, so we will have
guess the download status based on the daemon state, and be careful not
to tear down images too quickly.

We plan to present a simplified view to clients, based on the ticket
operations.

This patch prepares the operation class, attaching it to the ticket and
reporting the status will be done later.

Change-Id: Ic038350d3af6e09f83e374a84d35109119f4b3f5
Signed-off-by: Nir Soffer <nsoffer@redhat.com>