daemon: update timeout field in server.Tickets.get
This patch removes the timeout field from the Ticket object since it's
not useful as is (it's a constant that is received from the engine).
Instead, it sets the timeout field of the ticket info dictionary to be
the up to date timeout of the ticket.
This way the engine can use this field in the future for extending
tickets rather than managing a timeout mechanism itself.

Change-Id: Id1852cdf524af78b7439361581144f9fe86b41aa
Signed-off-by: Idan Shaby <ishaby@redhat.com>