daemon: enforce ticket's mandatory fields
This patch makes sure that when calling Tickets.put, the ticket
dictionary contains the mandatory size and ops fields.
Previously, authorizing a ticket with a missing size or ops keys would
have been failed with a generic KeyError. After this patch, it will fail
when trying to add the ticket with an InvalidTicket error.

Change-Id: I915bf39883a85be90f0f4697358b440402a1a731
Signed-off-by: Idan Shaby <ishaby@redhat.com>