Use simply expanded 'release_suffix'
Currently the Makefile defiles the 'release_suffix' using a rescursively
expanded variable:

  release_suffix = $(shell ...)

These variable will be expanded when it is used. If it is used multiple
times it will be evaluated multiple times. As the result contains a
date, that date can be different for different evaluations, which can
end up generating RPM files with different dates.

To avoid that this patch changes the makefile to use a simply expanded
variable, which is evaluated only once, when the variable is defined:

  release_suffix := $(shell ...)

Change-Id: I0fc48ac11bcf00a89c413698fffb40dc4702a82a
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>