proxy: Add the /tickets request handler
A client holding a signed ticket can register the ticket with the proxy
by performing a PUT request to /tickets. If the signed ticket is
verified the transfer ticket is stored in the proxy.

The stored ticket will be used later for verifying requests to
/images/ticket_id.

If the signed ticket is invalid, the request will fail with 403
Forbidden error. Previously we use 401 Unauthorized error, but in this
case we are not expecting or using an Authorization header.

See:
- https://tools.ietf.org/html/rfc7231#section-6.5.3
- https://tools.ietf.org/html/rfc7235#section-3.1

A client holding the ticket id can delete the ticket using a DELETE
request to /tickets/ticket_id.

Change-Id: I68008ae999a4b7d44206e60e2ee5722785efed9e
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Signed-off-by: Daniel Erez <derez@redhat.com>