Improve request logging
Log start message and finish/error message for each request. If request
failed, include the error in the log. Previously the error message was
hidden if there was no error handler that logged the error.

There is no need now to log errors if you raise an exception.

Here is an example of successful request:

2017-11-09 17:29:10,775 INFO    (Thread-29) [web] START [127.0.0.1] PUT /tickets/
2017-11-09 17:29:10,778 INFO    (Thread-29) [web] FINISH [127.0.0.1] PUT /tickets/: [200] 0 (0.00s)

And unsuccessful request:

2017-11-09 17:29:10,802 INFO    (Thread-32) [web] START [127.0.0.1] PUT /tickets/
2017-11-09 17:29:10,807 WARNING (Thread-32) [web] ERROR [127.0.0.1] PUT /tickets/:
[403] Error verifying signed ticket: Invalid ovirt ticket (data='this is
not a base64 encoded data', reason=Incorrect padding) (0.01s)

Change-Id: Iebc85130f7361fc5113dccb992392293a9b46f9e
Signed-off-by: Nir Soffer <nsoffer@redhat.com>