Add a failing test for http response version
For some reason we return HTTP/1.0 response, although we configure the
request handler to use HTTP/1.1. Needs more investigation.

Add failing tests for successful and error response, and enhance the
response() helper to print http version ("11" for HTTP/1.1, "10" for
HTTP/1.0).

When running the tests with -s, we see now:

(403, 'Forbidden', 10)
[('date', 'Sun, 12 Nov 2017 14:26:10 GMT'),
 ('content-length', '135'),
 ('content-type', 'application/json'),
 ('server', 'WSGIServer/0.1 Python/2.7.13')]

Change-Id: Ib40af171be6c336441b4f6d678daabbcb9d04ad0
Relates-To: https://bugzilla.redhat.com/1512317
Signed-off-by: Nir Soffer <nsoffer@redhat.com>