Add failing tests for checking actual image size
If image size is smaller than ticket size (not sure this can happen in
the current system, but it happened in 4.1), we validate the requested
size with the ticket size, and send a 200 OK response. Later we fail to
read the requested size, and fail with PartialContent exception.

On the client side the request fails with a timeout or IncompleteRead
after downloading gigabytes of data. The client cannot resume the
download to get the rest of the data, since we don't report the real
image size, only the ticket size.

There are 2 variants:

- Range request: the request should fail with 416 (Requested range not
  satisfiable)

- Complete request: the request should return the available data. The
  client does not care about the ticket, only about the existing image.

This patch adds a failing test for each variant.

Change-Id: I9898d2e5611eb246d1d37ed64e7111534ba8f82e
Relates-To: https://bugzilla.redhat.com/1512315
Signed-off-by: Nir Soffer <nsoffer@redhat.com>