Add failing test for invalid response
When downloading an empty image, we try to read ticket size bytes from
the image, and fail with PartialContent exception. For some reason the
WSGI server fail with:

Traceback (most recent call last):
  File "/usr/lib64/python2.7/SocketServer.py", line 596, in process_request_thread
    self.finish_request(request, client_address)
  File "/usr/lib64/python2.7/SocketServer.py", line 331, in finish_request
    self.RequestHandlerClass(request, client_address, self)
  File "/usr/lib64/python2.7/SocketServer.py", line 652, in __init__
    self.handle()
  File "/home/nsoffer/src/ovirt-imageio/daemon/ovirt_imageio_daemon/server.py", line 334, in handle
    handler.run(self.server.get_app())
  File "/usr/lib64/python2.7/wsgiref/handlers.py", line 92, in run
    self.close()
  File "/usr/lib64/python2.7/wsgiref/simple_server.py", line 33, in close
    self.status.split(' ',1)[0], self.bytes_sent
AttributeError: 'NoneType' object has no attribute 'split'

This looks like a python bug when an error occur before sending any
bytes, but the chances to get a fix for Python 2.7 are pretty low.

This will probably be fixed by checking the actual image size and
returning an empty response body.

This patch adds a failing test for this issue.

Change-Id: Ia10174c1c97d530e0f83aca0d77bdd7d519ce3c9
Relates-To: https://bugzilla.redhat.com/1512312
Signed-off-by: Nir Soffer <nsoffer@redhat.com>