daemon: catch PartialContent exception on upload
When the client pauses the upload, a PartialContent exception is raised
from directio.Receive.
Instead of logging the whole exception, we should log an error message
with the same content.

In the future, when we add an aborting mechanism, the daemon will not be
"surprised" in such cases and we won't have this error at all.

Change-Id: Ic238b7bd904f61c1ca482c0b5361ac2910235eba
Bug-Url: https://bugzilla.redhat.com/1523203
Signed-off-by: Idan Shaby <ishaby@redhat.com>