daemon: added options method to images
Describes the available methods for a resource. This is implemented
currently only for the ``/images`` resource. The request can be sent for
a concrete ticket-id, or the special ``*`` ticket_id.

The application can inspect the returned JSON document for available
features.

Available features:

- zero: PATCH/zero request is supported

Older versions did not implement this method, so the request would fail
with "405 Method Not Allowed".

Request:

    OPTIONS /images/%2A HTTP/1.1

Response:

    HTTP/1.1 200 OK
    Allow: GET,PUT,PATCH,OPTIONS
    Content-Type: application/json
    Content-Length: LENGTH

    {
        "features": ["zero"]
    }

Change-Id: Ibe5f4c09b0ce7a17125a09edbfa33b0efefb3e0a
Bug-Url: https://bugzilla.redhat.com/1560242
Signed-off-by: Daniel Erez <derez@redhat.com>