Document the random I/O APIs
We already support reading and writing byte ranges using GET and PUT,
but this support is optimized for transferring complete images.

We need to improve random I/O APIs for integration with virt-v2v and for
backup vendors. This patch documents the existing, changed, and new APIs
to be introduced in 1.3.

Change-Id: I98fbc20e1c05ab9fe627d33034034a47e1740da8
Bug-Url: https://bugzilla.redhat.com/1560242
Signed-off-by: Nir Soffer <nsoffer@redhat.com>