proxy: PATCH support
The proxy sends any PATCH /images/* requests to the daemon and return
the daemon response.

No validation is done on the content of the request except having some
content, so this patch adds support for zero, flush, and any future
methods supported by the daemon.

Since we expect an empty response, we send an empty response to the
caller. We will need more work if we decide to return non-empty response
in PATCH requests.

Change-Id: I95a8aacfece9f25a53ee969f92ebcfffe4a78d13
Bug-Url: https://bugzilla.redhat.com/1560242
Signed-off-by: Nir Soffer <nsoffer@redhat.com>