daemon: Fix OPTIONS to reflect actual capabilities
We assumed that a ticket with {ops: ["write"]} is also readable, but
this is not the case. The daemon checks for "read" op in GET, and for
"write" op in PUT and PATCH.

Turns out that upload tickets have only "write" op, so current code
report that we support GET, while issuing a GET request with fail with
"403 Forbidden".

Add a new test for a "write" only ticket, and modify the code to build
the correct method list, and hopefully be more clear.

Change-Id: I38a317e8aa69b6d94a788eabbe9dcceff410e8fb
Bug-Url: https://bugzilla.redhat.com/1560242
Signed-off-by: Nir Soffer <nsoffer@redhat.com>