proxy: Disable read_timeout for PATCH
We are using a read timeout to abort a request if the server does not
send a response after the specified timeout since request data was sent.

This does not work for PATCH, since the request is tiny, but the storage
operation may affect up to 2**64 bytes. There is no reasonable value
that we can use for such operations.

PATCH requests do not use read_timeout now.

Change-Id: I1edf31d0eb35a8101a9e14e866f494da8f04f6b0
Bug-Url: https://bugzilla.redhat.com/1560242
Signed-off-by: Nir Soffer <nsoffer@redhat.com>