daemon: Support "read" if having a "write" op
Until now we upload tickets were write only. It would be nice to add
"read" op if the user select this option in the SDK, but implementing
this is complicated and risky.

Change the policy so having "write" op imply also "read" op. With this
no change is need in the SDK or engine, and ticket progress will work
also for uploads initiated from the SDK (when implemented).

The downside is being able to read data uploaded to an image if you
could guess the random ticket id. Before this change, if you guessed the
ticket id you could only write data into the image. Since the ticket id
is random 128bit value, I don't think this is a real issue.

Change-Id: I859ddbf9b367e4aacaa276c3d8b0ec95e0af56c0
Signed-off-by: Nir Soffer <nsoffer@redhat.com>