proxy: Move SSL tests to new ssl_test module
We want to test a new configuration option for supporting legacy
TLSv1.1. For this we need to run the server both with default
configuration and with legacy configuration, so we cannot use
server_test module, reusing the same instance running with default
configuration.

Add a new ssl_test module, using a helper to create a server specific
configuration per test.

Change-Id: I469f29b0dddd30d0f4f13b58f7c39106f34fed58
Signed-off-by: Nir Soffer <nsoffer@redhat.com>