daemon: Move SSL tests to new ssl_test module
In server_test we use single server started using default configuration.
We want to test a server with different configuration to test TLSv1.1
support.

In the new module we create only the images service, since creating the
tickets service is not related to testing SSL. The service is created
using a context manager instead of pytest.fixture, so we can control the
configuration for each test.

Change-Id: I4d05f645a0a6d7c0f80f6679e7106dc016a44df9
Signed-off-by: Nir Soffer <nsoffer@redhat.com>