tests: Eliminate add_ticket and get_ticket helpers
Simplifying tickets.add() to accept a dict instead of a Ticket object
make it possible to use the same code to add a ticket during runtime and
during the tests, so add_ticket helper is not needed now.

get_ticket was trivial so it was inline into the callers, and tests that
can use the actual ticket use it directly instead of converting it to a
dict.

Change-Id: I723de8687243ba69a623181117e99ac2ece10b51
Signed-off-by: Nir Soffer <nsoffer@redhat.com>