daemon: Fix error handling with unknown ticket
When trying to download from unknown ticket and partial or no range,
fail with "403 Forbidden" instead of "500 Internal Server Error" and a
useless traceback.

This broke when we add support for partial content, because we had to
get the ticket to satisfy ticket.authorize, and we forgot to handle
KeyError when a ticket does not exists. Fixed by changing authorize to
accept offset and size arguments, so we don't need to get the ticket
before authorization.

Also add a missing test for the case of empty ticket id.

This change fix 2 expected failures in local_test module. One of the
test was wrong, sending invalid patch message. After fixing the test it
passes now.

Change-Id: Idb147dfd365ee0f1feb5b2b458baca6149045573
Signed-off-by: Nir Soffer <nsoffer@redhat.com>