uhttp: Support abstract sockets
Abstract socket are not bound to filesystem pathname, and do not use
permissions or ownership. We like to use an abstract socket for
accessing images locally.

Skip unlinking of abstract socket (starting with "\0"), since there is
noting on the filesystem to remove, and os.unlink() does not like
pathname with embedded nulls.

Using empty socket address will bind the server to random abstract
socket picked by the kernel. This is useful for testing on a host with a
running daemon.

Change-Id: Idbe4d6ef7eb8cb8c5864bc8682c4fa2a1e011161
Signed-off-by: Nir Soffer <nsoffer@redhat.com>