config: Improve default config documentation
The daemon configuration file does not need to specify any default
value. All defaults are set using the config module. The purpose of the
configuration file is to document the possible options and explain why
the default values are chosen, helping the user to change the values,
and restore the original value if needed.

Improve the documentation both in the config module and in the
configuration file, and comment the default value in the config file.

Change-Id: Ie58cd1f881d8ec7779d95f9dd8a732782670f6b1
Signed-off-by: Nir Soffer <nsoffer@redhat.com>