profile: Add CPU profiling
If yappi is installed on a host, we can create a cpu profile of the
daemon.

To start profiling, use:

    curl --unix-socket /run/vdsm/ovirt-imageio-daemon.sock \
        -X POST http://localhost/profile?run=y

This uses the "cpu" clock, measuring CPU time spent in every function. To
measure wall clock time, use:

    curl --unix-socket /run/vdsm/ovirt-imageio-daemon.sock \
        -X POST http://localhost/profile?run=y&clock=wall

To stop profiling, use:

    curl --unix-socket /run/vdsm/ovirt-imageio-daemon.sock \
        -X POST http://localhost/profile?run=n

Stopping profiling writes the collected profile to
/tmp/ovirt-imageio-daemon.prof. This path can changed using the
configuration file.

Change-Id: If1be070d85dbc606fc8a1e926e827626381dd586
Signed-off-by: Nir Soffer <nsoffer@redhat.com>