util: Display microseconds in repr(util.Clock)
We used to display 10 milliseconds resolution, since we used to time
complete upload operation and that takes seconds. However with virt-v2v,
we have many small requests, and we want to focus on making them fast.
This means we need sub millisecond resolution. Change __repr__ to show
microseconds.

Here is an example of new clock log:

2018-06-13 21:55:53,756 INFO    (Thread-808) [directio] Operation stats:
<Clock(total=0.007067, read=0.001220, write=0.005244)>

Change-Id: If3ff8a8d555334b44dad6887e1234f61427b9442
Signed-off-by: Nir Soffer <nsoffer@redhat.com>