web: Record timing info for entire request
Previously we recorded the time spent during directio.Operation()
subclasses. This was very useful to improve the performance of the
operations, but turns out we have performance issue with small requests.

The web.Application framework creates now a clock for timing the entire
request. The clock is started when a request is received, and passed
down to the request handler to time handler specific operations.

Previously we log the recorded time when an operation was completed, and
log the total request time in web.Application. However during download
the FINISH log message was logged before the operation was completed,
because of the twisted way WSGI works.

Now the FINISH log message is logged when the WSGI app_iter is closed,
and the "Operation stats" timings are logged in the "FINISH" log, so
simple grep reveal the timing info for all requests.

Here are example FINISH log for PUT and GET requests:

2018-06-14 02:10:16,162 INFO    (Thread-51) [web] FINISH [127.0.0.1] PUT
/images/0841f894-5331-4b4a-b473-2499e6eb0f3e: [200] 0 [request=0.005057,
operation=0.002957, read=0.000015, write=0.000023, sync=0.002575]

2018-06-14 02:10:16,303 INFO    (Thread-64) [web] FINISH [127.0.0.1] GET
/images/77122a6a-37bf-4d29-b8ab-8e47a75b66f6: [206] 1023
[request=0.001747, operation=0.000945, read=0.000277]

All request handler classes accept now an optional clock argument. We
will use it later to time other areas of the application as needed.

Change-Id: I36dbcc5325afce99faa564c6e3217e552fae60fc
Signed-off-by: Nir Soffer <nsoffer@redhat.com>