tests: Fix flaky download progress test
The tests assumed that if we read all the data, the server has completed
the request and marked the ticket as active. However it is possible that
the server as completed the write but the kernel switched to the test
process to perform the assert before the server could run again.

Until we find a better way to test this, add little sleep to make sure
the server process will run again before the test assert about the
ticket state.

Change-Id: Iaddf02705cf59d216bdea10f0039ddfe3952a3cf
Signed-off-by: Nir Soffer <nsoffer@redhat.com>