wsgi: Close connection on errors
If a request with a body fail before the body was read from the socket,
for example because a ticket expired, we send a response without reading
the body. This is not very nice and compliant with the HTTP protocol,
but we really have no choice, since the body may be 100 GiB of data.

Since there is no way to recover from such error, the only behavior that
make sense is to close the connection. Detect this situation in the
server and close the connection.

Note that in some cases we know that the request body was read, and we
don't need to close the connection, but we need to find a good way to
signal this condition from the application code. Even if we find a good
way, it is very fragile and easy to get wrong.

The old tests for keeping connections and closing them were completely
broken. Replace them with proper failing tests, since WSGI does not
support keep alive connection.

Change-Id: Idce6d232cd7625d25a06ce18d96110c2322e4e38
Signed-off-by: Nir Soffer <nsoffer@redhat.com>