web: Always set response content length
In commit ba5c3af8be0e (web: Record timing info for entire request) we
wrapped all responses with a LoggingAppIter, logging the FINISH line for
every request. This change introduced unintended behavior change, not
returning a Content-Length for "204 No Content" response. Before this
change, WSGI and webob magic were adding a Content-Length magically.

That change broke delete ticket flow:

- vdsm deletes a ticket
- daemon deletes a ticket and returns "204 No Content"
- vdsm fails because Content-Length is missing
- engine retry
- vdsm deletes a ticket
- daemon fails with "404 Not Found"
- vdsm passes failure to engine
- engine retry forever - endless loop!

Turns out that the current behaviour is correct according the HTTP 1.1
RFC[1]:

    A server MUST NOT send a Content-Length header field in any response
    with a status code of 1xx (Informational) or 204 (No Content).

But we cannot break old clients by this change, so the behavior change
was reverted, and we always return Content-Length. I don't think there
is any client that may break because of this change, since the client
should not look at headers that the server must not send.

[1] https://tools.ietf.org/html/rfc7230#section-3.3.2

Change-Id: I5839fc66c542a6d7f8820063440506b7e115ee3b
Signed-off-by: Nir Soffer <nsoffer@redhat.com>