daemon: Make DELETE idempotent
We used to fail with "404 Not Found" when trying to delete a
non-existing ticket. This is generally a good approach to reveal bugs in
clients, but it make the API harder to use considering failures in
distributed environment.

This also causes endless loop if deleting a ticket succeeded, but vdsm
failed to process the successful response, or fail to send a reply to
engine.

Change delete to be idempotent, so we can call it multiple times and get
the same response "204 No Content". This is also the way it should be
implemented according to the RFC[1].

[1] https://tools.ietf.org/html/rfc7231#section-4.2.2

Change-Id: I3e87e9a187d45c614e394aad7180141260306fc7
Signed-off-by: Nir Soffer <nsoffer@redhat.com>