tests: Add fallocate test script
Rename tests/zeroout.py to test/fallocate.py, and modify it to work with
both block device and files. This tests now both directio.FileIO() when
zeroing a file, and directio.BlockIO() when zeroing a block device.

Note: unlike the fallocate command, we always allocate space when
zeroing a file. We don't support yet sparse images.

Change-Id: Ifddab13b7cf35dd9c331c5bb9cf24238fcf2e5a2
Bug-Url: https://bugzilla.redhat.com/1615144
Signed-off-by: Nir Soffer <nsoffer@redhat.com>