config: Add option to disable the local service
This will allow to disable the local service when installing on engine
host as a proxy. Add a config option to disable it. This may also be
handy for testing, or for disabling this feature temporarily if this
feature cause trouble in the field.

When the feature is disabled, we do not report the "unix_socket" feature
in OPTIONS.

This patch is base on Nir's patch
https://gerrit.ovirt.org/c/93001

Change-Id: I519161913907aa3a4032957c35606f1bd989c89d
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>