web: Avoid null bytes in web logs
webob.Request.client_addr is returning the insecure value of
"X-Forwarded-Host" or the first item of "REMOTE_ADDR". When the client
connects via a unix socket, REMOTE_ADDR is set to the first character of
the unix socket path; "/" for the control socket, and "\0" for the local
service using abstract unix domain socket. Both "/" and "\0" are not
useful, and "\0" also causes trouble when grepping the logs:

$ grep web /var/log/ovirt-imageio-daemon/daemon.log
Binary file /var/log/ovirt-imageio-daemon/daemon.log matches

Use REMOTE_HOST if set, or REMOTE_ADDR. On the unix socket based server
we always get "localhost", and in the HTTPS based server we get the
client IP address (e.g. 1.2.3.4).

Here are example logs with this change:

2018-07-13 18:29:14,161 INFO    (Thread-14) [web] START: [localhost] GET
/images/118cc1bd-06e4-403b-939b-9654cb6489d8

2018-07-13 18:29:14,162 INFO    (Thread-14) [web] FINISH [localhost] GET
/images/118cc1bd-06e4-403b-939b-9654cb6489d8: [200] 1024
[request=0.001385, operation=0.000859, read=0.000281]

Change-Id: I1e624256c6852ea29f5c1ad1529789935e59d4ef
Signed-off-by: Nir Soffer <nsoffer@redhat.com>